# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: KAX - Kannaka Artifact Exchange API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /artifacts:
    get:
      operationId: listArtifacts
      tags: [artifacts]
      summary: List all artifacts
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [raw, scored, narrated, dropped]
        - name: minScore
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: search
          in: query
          schema:
            type: string
        - name: artifactType
          in: query
          schema:
            type: string
            enum: [image, audio, music, text, furniture]
        - name: editionType
          in: query
          schema:
            type: string
            enum: [open, limited, "1_of_1"]
        - name: agentId
          in: query
          schema:
            type: integer
        - name: all
          in: query
          description: Admin only. When `true`, returns data across all owners; otherwise scoped to the caller.
          schema:
            type: boolean
      responses:
        "200":
          description: List of artifacts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactListResponse"
  /artifacts/{id}:
    get:
      operationId: getArtifact
      tags: [artifacts]
      summary: Get artifact by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Artifact details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
  /artifacts/{id}/score:
    post:
      operationId: scoreArtifact
      tags: [artifacts]
      summary: Run taste engine scoring on an artifact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scored artifact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
  /artifacts/{id}/narrate:
    post:
      operationId: narrateArtifact
      tags: [artifacts]
      summary: Generate narrative for an artifact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Narrated artifact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
...
components:
...
  schemas:
...
    Artifact:
      type: object
      properties:
        id:
          type: integer
        externalId:
          type: string
        title:
          type: string
        creatorName:
          type: string
        publicUrl:
          type: string
        thumbnailUrl:
          type: string
          nullable: true
        reactionCount:
          type: integer
        artifactType:
          type: string
          enum: [image, music, text, audio, furniture]
        status:
          type: string
          enum: [raw, scored, narrated, dropped]
        kannakaScore:
          type: number
          nullable: true
        rarityScore:
          type: number
          nullable: true
        narrative:
          type: string
          nullable: true
        narrativeTitle:
          type: string
          nullable: true
        transmissionId:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        ingestedAt:
          type: string
          format: date-time
        scoredAt:
          type: string
          format: date-time
          nullable: true
        narratedAt:
          type: string
          format: date-time
          nullable: true
        dropId:
          type: integer
          nullable: true
        ownerId:
          type: string
          nullable: true
        obcArtifactUuid:
          type: string
          nullable: true
        agentId:
          type: integer
          nullable: true
        editionType:
          type: string
          enum: [open, limited, "1_of_1"]
        editionTotal:
          type: integer
          nullable: true
        editionSerial:
          type: integer
          nullable: true
        heat:
          type: integer
        lastReactionAt:
          type: string
          format: date-time
          nullable: true
        scoreBreakdown:
          type: object
          nullable: true
          properties:
            reactionSignal:
              type: number
            heatSignal:
              type: number
            novelty:
              type: number
            exploration:
              type: number
            baseScore:
              type: number
            scarcityMultiplier:
              type: number
            editionType:
              type: string
            finalScore:
              type: number
          required:
            - reactionSignal
            - novelty
            - exploration
            - baseScore
            - scarcityMultiplier
            - editionType
            - finalScore
      required:
        - id
        - externalId
        - title
        - creatorName
        - publicUrl
        - reactionCount
        - artifactType
        - status
        - tags
        - ingestedAt
        - editionType
...
    ArtifactListResponse:
      type: object
      properties:
        artifacts:
          type: array
          items:
            $ref: "#/components/schemas/Artifact"
        total:
          type: integer
      required:
        - artifacts
        - total
...