# lib/api-zod/src/generated/types/artifact.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KAX - Kannaka Artifact Exchange API
 * OpenAPI spec version: 0.1.0
 */
import type { ArtifactArtifactType } from "./artifactArtifactType";
import type { ArtifactEditionType } from "./artifactEditionType";
import type { ArtifactScoreBreakdown } from "./artifactScoreBreakdown";
import type { ArtifactStatus } from "./artifactStatus";

export interface Artifact {
  id: number;
  externalId: string;
  title: string;
  creatorName: string;
  publicUrl: string;
  thumbnailUrl?: string | null;
  reactionCount: number;
  artifactType: ArtifactArtifactType;
  status: ArtifactStatus;
  kannakaScore?: number | null;
  rarityScore?: number | null;
  narrative?: string | null;
  narrativeTitle?: string | null;
  transmissionId?: string | null;
  tags: string[];
  ingestedAt: Date;
  scoredAt?: Date | null;
  narratedAt?: Date | null;
  dropId?: number | null;
  ownerId?: string | null;
  obcArtifactUuid?: string | null;
  agentId?: number | null;
  editionType: ArtifactEditionType;
  editionTotal?: number | null;
  editionSerial?: number | null;
  heat?: number;
  lastReactionAt?: Date | null;
  scoreBreakdown?: ArtifactScoreBreakdown;
}